pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ownmail"
version = "0.2.0-dev"
description = "Own your mail. A file-based email backup and search tool."
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "Chungmin Lee" }
]
keywords = ["email", "backup", "gmail", "imap", "archive", "search", "self-hosted"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Topic :: System :: Archiving :: Backup",
]
dependencies = [
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "keyring>=24.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.scripts]
ownmail = "ownmail.cli:main"

[project.urls]
Homepage = "https://github.com/chungmin/ownmail"
Repository = "https://github.com/chungmin/ownmail"
Issues = "https://github.com/chungmin/ownmail/issues"

[tool.hatch.build.targets.wheel]
packages = ["ownmail"]

[tool.hatch.build.targets.sdist]
include = [
    "ownmail/",
    "README.md",
    "LICENSE",
]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

# Ruff linter configuration
[tool.ruff]
target-version = "py38"
line-length = 120
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict= (Python 3.10+)
]

[tool.ruff.lint.isort]
known-first-party = ["ownmail"]

# Coverage configuration
[tool.coverage.run]
source = ["ownmail"]
omit = ["tests/*", "venv/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
